requirements.txt
flask
werkzeug
transformers
torch
torchvision
Pillow
gradio
requests
fastapi
uvicorn
timm
mediapipe
PyJWT
opencv-python
fastai
scikit-learn
numpy
pandas